+++ contracts/lending/interfaces/ISecuredLoanReceiver.sol
@@ -1,7 +1,12 @@
 pragma solidity >=0.5.12 <=0.6.2;
 
-import { IBPool } from "./IBPool.sol";
+import { IBPool } from "../../pool/interfaces/IBPool.sol";
 
 interface ISecuredLoanReceiver {
-    function secureLoan(IBPool optionPool, uint loanedQuantity, uint fee, bytes calldata params) external returns(bool);
+    function secureLoan(
+        IBPool optionPool,
+        uint256 loanedQuantity,
+        uint256 fee,
+        bytes calldata params
+    ) external returns (bool);
 }
